requirements.txt
python-telegram-bot[webhooks]>=20.0
Pillow>=10.0.0
python-dotenv>=1.0.0
